ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
nltk==3.8.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
psutil==5.9.6
python-multipart==0.0.6